update assetstools for ulong texture offset

+++ TexturePlugin/Program.cs
@@ -39,7 +39,7 @@ namespace TexturePlugin
                 if (File.Exists(fixedStreamPath))
                 {
                     Stream stream = File.OpenRead(fixedStreamPath);
-                    stream.Position = texFile.m_StreamData.offset;
+                    stream.Position = (long)texFile.m_StreamData.offset;
                     texFile.pictureData = new byte[texFile.m_StreamData.size];
                     stream.Read(texFile.pictureData, 0, (int)texFile.m_StreamData.size);
                 }
@@ -222,7 +222,7 @@ namespace TexturePlugin
                     AssetBundleDirectoryInfo06 info = dirInf[i];
                     if (info.name == searchPath)
                     {
-                        reader.Position = bundle.bundleHeader6.GetFileDataOffset() + info.offset + streamInfo.offset;
+                        reader.Position = bundle.bundleHeader6.GetFileDataOffset() + info.offset + (long)streamInfo.offset;
                         texFile.pictureData = reader.ReadBytes((int)streamInfo.size);
                         texFile.m_StreamData.offset = 0;
                         texFile.m_StreamData.size = 0;
